+++ ch6/keyview2/keyview2.go
@@ -88,8 +88,8 @@ func wndproc(hwnd win32.HWND, msg uint32, wParam, lParam uintptr) (result uintpt
 
 	case win32.WM_SIZE:
 		if msg == win32.WM_SIZE {
-			cxClient = int32(win32.LOWORD(lParam))
-			cyClient = int32(win32.HIWORD(lParam))
+			cxClient = win32.LOWORD(lParam)
+			cyClient = win32.HIWORD(lParam)
 		}
 
 		// Calculate scrolling rectangle.
